refactor(bridge): move OutboundMessage from pool/ to actor/
OutboundMessage defines the message types consumed by the writer actor,
so it belongs in the actor module alongside writer.rs and response_router.rs.

This improves module cohesion:
- actor/: defines what the actor processes (OutboundMessage)
- pool/: manages connections and uses actor APIs

Import paths updated in:
- actor/writer.rs: use super::outbound_message
- pool/connection_handle.rs: import from actor
- pool/message_sender.rs: import from actor
- pool.rs tests: import from actor
